requirements.txt
numpy
pandas
plotly
matplotlib
tensorflow-gpu